+++ packages/react-storage/src/components/index.ts
@@ -1,3 +1,5 @@
+export { StorageImage, StorageImageProps } from './StorageImage';
+
 export {
   StorageManager,
   StorageManagerProps,