+++ packages/react-storage/src/index.ts
@@ -1,4 +1,6 @@
 export {
+  StorageImage,
+  StorageImageProps,
   StorageManager,
   StorageManagerProps,
   DropZoneProps,