+++ packages/react/src/primitives/types/heading.ts
@@ -5,6 +5,7 @@ export type HeadingLevel = 1 | 2 | 3 | 4 | 5 | 6;
 
 export type HeadingTag = 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6';
 
+/** @deprecated For internal use only */
 export interface BaseHeadingProps extends BaseTextProps {
   /**
    * @description