+++ packages/react/src/primitives/types/passwordField.ts
@@ -4,6 +4,7 @@ import { BaseTextFieldProps } from './textField';
 import { ElementType, PrimitiveProps } from './view';
 import { BaseInputProps } from './input';
 
+/** @deprecated For internal use only */
 export interface BasePasswordFieldProps extends BaseTextFieldProps {
   /**
    * @description
@@ -66,6 +67,7 @@ export type PasswordFieldProps<Element extends ElementType = 'input'> =
 
 export type PasswordFieldType = 'password' | 'text';
 
+/** @deprecated For internal use only */
 export interface BaseShowPasswordButtonProps
   extends BaseButtonProps,
     Pick<BaseInputProps, 'hasError'> {